pyproject.toml
[tool.poetry]
name = "adopet"
version = "0.1.0"
description = ""
authors = ["derek.coelho <derek.coelho.ext@vr.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.10.3"
fastapi = "^0.95.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
aio-pika = "^9.0.5"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.9"}
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
requests = "^2.28.2"
sqlalchemy-utils = "^0.41.0"
phonenumbers = "^8.13.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
uvicorn = "^0.21.1"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
asgi-correlation-id = "^4.2.0"
pyyaml = "^6.0"
faker = "^18.6.2"
httpx = "^0.24.0"
anyio = "^3.6.2"
pytest-freezegun = "^0.4.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
python-dotenv = "^1.0.0"
pytest-mock = "^3.10.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
